refactor getIdentityProtocol to use it in draft validation
