+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- More permissive protocol validation in draft according to identities
+
 ## [0.13.2] 2018-12-11
 
 ### Added
